SidebarComponentをリファクタリング #25
