+++ client/src/components/Dashboard/Dashboard.js
@@ -1,17 +1,30 @@
-import React, { useEffect } from "react";
+import React, { useEffect, useState } from "react";
+import SignedinNavbar from "../Layout/SignedinNavbar";
 import Sidebar from "./Sidebar";
 import CreateWorkspace from "../../containers/Workspace/CreateWorkspace";
 
 const Dashboard = ({ workspaces, isLoading, getWorkspaces }) => {
+  const [open, setOpen] = useState(false);
+
+  const handleDrawerOpen = () => {
+    setOpen(true);
+  };
+
+  const handleDrawerClose = () => {
+    setOpen(false);
+  };
+
   useEffect(() => {
     getWorkspaces();
   }, []);
+
   if (isLoading) {
     return <div>isLoading</div>;
   }
   return (
     <div>
-      <Sidebar />
+      <SignedinNavbar open={open} handleDrawerOpen={handleDrawerOpen} />
+      <Sidebar open={open} handleDrawerClose={handleDrawerClose} />
       <div className="none-workspace">
         <div className="workspace-description">
           <p>まだworkspaceがありません。</p>
